Use full path to api in stead of relatieve path
At this moment a relative path is used (../../../), for the cde saikuWidget. But the preview of the dashboard and the 'normal' view are on a different directory level, so one of them needs an extra ../ Fixed by using the full path /pentaho/plugin/saiku/api But this won't work if somebody use a different 'directory' for pentaho. But I guess that is not a big problem...